Add a summarization message type. (#110)
This will be used to convey to the client that summarization is occurring/has occurred. For now this will be displayed primary by a change from warping... to summarizing context..., perhaps with a timer. Eventually this could also be used to convey the content of summaries should we choose to expose that to the user.