Add a SummarizeConversation SystemQuery variant (#150)
* Add a SummarizeConversation SystemQuery variant that acts as the input message indicator that the user has requested summarization.

Note that the SystemQuery type is already somewhat broaded by the inclusion of other user activated requests such as CreateNewProject, CloneRepository, and ResumeConversation. At this point the type is more broadly a category of requests from the client that are not strictly speaking UserQueries. So I'm including SummarizationConversation here for consistency.

* Regen