Add run command tool call ID param to CLI subagent query. (#160)
## Description

This PR enables a fix for a subtle bug caught by maggie in https://github.com/warpdotdev/warp-internal/pull/19391#issuecomment-3537272082 -- specifically, the dangling tool call issue. In reality, its not a missing tool call result, but a tool call and result with a user message spliced in between.

This state is specifically reproducible with LRC X Inline AI enabled, in a corner case where the agent has run a command, but the user immediately sends a follow up query _after_ the command starts execution but _before_ the client actions to spawn the subagent have not yet been received/handled

In this scenario, we treat the user query like an Inline AI x LRC query- which is correct - its a query to spawn a subagent on a running command, which in this case just happens to be a requested command.  And in fact, this makes it possible for the client to send an agent request _any_ time, which is desirable.

However, the inline ai/LRC codepath currently assumes the running command is always user-executed, which in this specific edge case is not true. Then, later, when the subagent is finished, we check whether or not the original subagent tool call was spawned for an agent requested command to determine if we should format a summary of the interaction or if we should format a run command tool call result.

This change allows us to format the manual/inline ai-spawned subagent interaction the same way we would have formatted the automatically spawned subagent if the running command is a requested command, which ends up fixing the tool call result error.