Differentiate token usage origins (#170)
## Description

This PR adds a separate mapping for token usage by category in the ConversationUsageMetadata message.

Note: the server-side implementation of this proto is here: https://github.com/warpdotdev/warp-server/pull/7480

## Justification

We consistently get feedback from folks who are confused why models that they haven't selected are showing up in the model picker. To fix this, we want to indicate what different models are being used for (i.e. compaction, cli commands, summarization, etc.). To effectively do this in the UI, we need to delineate token usage by category, as `warp_token_usage` & `byok_token_usage` are the maps we use to generate the model list in the usage footer.